# docs/swagger.yaml
definitions:
...
  models.Product:
    properties:
      availability:
        type: string
      category:
        type: string
      createdAt:
        type: string
      depth:
        type: number
      description:
        type: string
      height:
        type: number
      id:
        type: integer
      image:
        type: string
      price:
        type: number
      rating:
        type: number
      returnPolicy:
        type: string
      shipping:
        type: string
      stock:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      warranty:
        type: string
      weight:
        type: integer
      width:
        type: number
    type: object
  models.ProductReview:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      productId:
        type: integer
      rating:
        type: number
      review:
        type: string
      reviewBy:
        type: integer
      updatedAt:
        type: string
    type: object
...
  service.AllReviewsResponse:
    properties:
      customer:
        $ref: '#/definitions/service.Customer'
      review:
        $ref: '#/definitions/service.Review'
    type: object
  service.Customer:
    properties:
      age:
        type: string
      avatar:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      middlename:
        type: string
    type: object
...
  service.ProductReviewPayload:
    properties:
      rating:
        type: number
      review:
        type: string
    required:
    - rating
    - review
    type: object
  service.ProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      totalCount:
        type: integer
    type: object
...
  service.Review:
    properties:
      createdAt:
        type: string
      id:
        type: string
      rating:
        type: number
      review:
        type: string
      updatedAt:
        type: string
    type: object
...
  utils.ResUser:
    properties:
      message:
        type: string
    type: object
  utils.ResUserWithData:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: prachurjyagogoi123@gmail.com
  description: This is a dummy backend for an ecommerce store.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce Backend API
  version: "1.0"
paths:
...
  /favorite/product/{productId}:
    delete:
      description: API for removing a product as favorite.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: UnMark a product as favorite.
      tags:
      - Product
    post:
      description: API for making a product as favorite.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Mark a product favorite.
      tags:
      - Product
  /favorite/products:
    get:
      description: API for getting all favorite products.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get all favorite products.
      tags:
      - Product
...
  /product/{productId}:
    get:
      description: API for fetching a product.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get a product details.
      tags:
      - Product
  /product/{productId}/reviews:
    get:
      description: API for fetching all the reviews for a product made by multiple
        users.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AllReviewsResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get all reviews for a product.
      tags:
      - Product
  /products:
    get:
      description: API for fetching all products.
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: The number of products to fetch
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  $ref: '#/definitions/service.ProductsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get all products.
      tags:
      - Product
...
  /review/product/{productId}:
    post:
      consumes:
      - application/json
      description: API for adding a product review by user.
      parameters:
      - description: Product review payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.ProductReviewPayload'
      - description: Product id
        in: path
        name: productId
        required: true
        type: string
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductReview'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Add a product review.
      tags:
      - Product
  /review/product/{reviewId}:
    delete:
      consumes:
      - application/json
      description: API for deleting a product review by user.
      parameters:
      - description: Review ID
        in: path
        name: reviewId
        required: true
        type: string
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductReview'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Delete a product review.
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: API for updating a product review by user.
      parameters:
      - description: Product review payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.ProductReviewPayload'
      - description: Review ID
        in: path
        name: reviewId
        required: true
        type: string
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductReview'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Update a product review.
      tags:
      - Product
...
swagger: "2.0"